refactor: dto 변환 로직 presentation 레이어로 이동
