+++ src/main/java/roomescape/presentation/dto/request/PaymentProcessRequest.java
@@ -1,0 +1,4 @@
+package roomescape.presentation.dto.request;
+
+public record PaymentProcessRequest(String paymentKey, String orderId, String amount) {
+}
